# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LuizaLabs - Desafio técnico - Vertical Logística",
    "version": "1.0.0"
  },
  "paths": {
    "/orders": {
      "get": {
        "summary": "Retorna todos os pedidos ou pedidos entre duas datas",
        "description": "Pode ser usado com ou sem os parâmetros de data. Se `dateStart` e `dateEnd` forem fornecidos, retorna apenas os pedidos nesse intervalo.",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dateStart",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 20210303
            },
            "description": "Data inicial no formato YYYYMMDD"
          },
          {
            "in": "query",
            "name": "dateEnd",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 20210305
            },
            "description": "Data final no formato YYYYMMDD"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user_id": {
                        "type": "integer"
                      },
                      "user_name": {
                        "type": "string"
                      },
                      "order_id": {
                        "type": "integer"
                      },
                      "product_id": {
                        "type": "integer"
                      },
                      "product_price": {
                        "type": "number"
                      },
                      "order_date": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar pedidos"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Buscar pedido por ID",
        "description": "Retorna os dados de um pedido específico com base no ID fornecido.",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do pedido"
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user_id": {
                        "type": "integer"
                      },
                      "user_name": {
                        "type": "string"
                      },
                      "order_id": {
                        "type": "integer"
                      },
                      "product_id": {
                        "type": "integer"
                      },
                      "product_price": {
                        "type": "number"
                      },
                      "order_date": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Pedido não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Pedido não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar o pedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar pedido"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/upload": {
      "post": {
        "summary": "Enviar arquivo de pedidos",
        "description": "Faz o upload de um arquivo contendo pedidos para serem processados.",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo contendo os pedidos"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Arquivo salvo com sucesso. Dados em processo de importação"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Arquivo ausente ou inválido"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao processar upload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro no upload"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}